Fixed errornous getter of the gyro and magnetic field sensor which caused cassandra to fail on mapping data

+++ AssistancePlatformModuleAPI/src/de/tudarmstadt/informatik/tk/assistanceplatform/data/sensor/Gyroscope.java
@@ -88,7 +88,7 @@ public class Gyroscope extends SensorData {
 	}
 
 
-	public double getXUncalibratedNoDriftOptional() {
+	public float getXUncalibratedNoDriftOptional() {
 		return xUncalibratedNoDriftOptional;
 	}
